web gabut/backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
passlib[bcrypt]
python-jose[cryptography]